+++ endpoints/meals.py
@@ -3,8 +3,6 @@ from models import db, Meal, MealCategory, Diet, MealIngredients, MealHistory, U
 from datetime import datetime
 from sqlalchemy import or_
 
-# Pobieranie posiłków
-
 def get_meals():
     limit = request.args.get('limit', default=10, type=int)
     page = request.args.get('page', default=1, type=int)
@@ -23,8 +21,11 @@ def get_meals():
     })
 
 def get_meal(meal_id):
-    meal = Meal.query.get_or_404(meal_id)
-    return jsonify(meal.to_dict())
+    meal = Meal.query.get(meal_id)
+    if meal:
+        return jsonify(meal.to_dict())
+    else:
+        return jsonify({"message": "Meal not found"}), 404
 
 def search_meals():
     query = request.args.get('query', default='', type=str)
@@ -65,7 +66,6 @@ def search_meals():
 def create_meal():
     data = request.get_json()
 
-    # Validate required fields
     if not data.get('name') or not data.get('creator_id'):
         return jsonify({"error": "Name and creator_id are required"}), 400
     
@@ -73,14 +73,12 @@ def create_meal():
     # My bierzemy je z body dla celów prezentacyjnych
     creator_id = data.get('creator_id')
 
-    # Validate category_id
     category_id = data.get('category_id')
     if category_id:
         category = MealCategory.query.get(category_id)
         if not category:
             return jsonify({"error": "Category not found"}), 404
 
-    # Validate diet_id
     diet_id = data.get('diet_id')
     if diet_id:
         diet = Diet.query.get(diet_id)
@@ -122,10 +120,9 @@ def update_meal(meal_id):
     if data.get('ingredients'):
         return jsonify({"error": "To update ingredients, use the /meals/<meal_id>/ingredients endpoint"}), 400
 
-    # Fetch the meal to be updated
-    meal = Meal.query.get_or_404(meal_id)
-
-    # Save old version to MealHistory
+    meal = Meal.query.get(meal_id)
+    if not meal:
+        return jsonify({"message": "Meal not found"}), 404
     meal.save_meal_version()
 
     # Update meal details
@@ -143,7 +140,9 @@ def update_meal(meal_id):
     return jsonify({"message": "Meal updated", "meal_id": meal.id}), 200
 
 def delete_meal(meal_id):
-    meal = Meal.query.get_or_404(meal_id)
+    meal = Meal.query.get(meal_id)
+    if not meal:
+        return jsonify({"message": "Meal not found"}), 404
     db.session.delete(meal)
     db.session.commit()
     return jsonify({"message": "Meal deleted"})
